Adds Ajaxy goodness
Brand new class .api-form, which will dynamically build an Ajax request from the form elements and send it to the forms actions.

 Brand new toast system, thanks to the toastr library.

Speaking of toasts, if you set the data-toast attribute on an form.api-form it will now use that as the "in process" toast. As for success and failure toasts, those should be returned by the new build_api_response function.